begin navbar updates
